+++ ESP32_GuitarCable/src/FIFO.cpp
@@ -48,6 +48,6 @@ uint8_t pop_Sample(uint16_t *sample) {
 
 
 // Check how Many Samples are Pre-Buffered
-uint16_t sampleCount() {
+uint32_t sampleCount() {
     return (PREBUFF_LEN + head - tail) % PREBUFF_LEN;
 }
